+++ example/functions.js
