+++ example/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "substrate-perf-test-app",
+  "name": "substrate-perf-test-example",
   "version": "1.0.0",
   "description": "Rpc perf app using artillery",
   "main": "index.js",
