+++ package.json
@@ -11,7 +11,7 @@
     "url": "git+https://github.com/dwellir-public/artillery-engine-substrate.git"
   },
   "author": "Ankan <me@theduck.dev>",
-  "license": "ISC",
+  "license": "Apache-2.0", 
   "dependencies": {
     "@polkadot/api": "^8.7.1",
     "artillery-core": "^4.8.1"
